// microservicios-web/src/app/views/dashboard/Setup/product/models/product.ts
export class Product {
  id?: number
  nombreProducto?: string;
  marca?: string;
  imei?: string;
  color?: string;
  descripcion?: string;
  precio?: number;
  stockDisponible?: number;
  numeroSerie?: string;
  especificacionesTecnicas?: string;
  estadoProducto?: string;
  categoriaTipo?: string;
  fechaIngreso?: null;





}